♻️ Refactor: 분리되어있던 query option 통합
